pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-router"
version = "2.0.0"
description = "AI-powered email processing and routing system with multi-provider support"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Email Router Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    
    # Google Cloud
    "functions-framework>=3.5.0",
    "google-cloud-pubsub>=2.18.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "google-generativeai>=0.3.0",
    
    # Email Providers
    "mailgun>=0.1.0",
    
    # LLM Providers
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "langchain>=0.1.0",
    "litellm>=1.0.0",
    
    # Memory/Vector Stores
    "redis[hiredis]>=5.0.0",
    "pinecone-client>=2.2.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    
    # Multi-Agent Frameworks
    "langgraph>=0.0.20",
    "crewai>=0.1.0",
    
    # Utilities
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    
    # Async Support
    "asyncio>=3.4.3",
    "aiofiles>=23.2.0",
    "aiocache>=0.12.0",
    
    # Security
    "cryptography>=41.0.0",
    "python-jose[cryptography]>=3.3.0",
    
    # Monitoring
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

ml = [
    "scikit-learn>=1.3.0",
    "numpy>=1.25.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "wandb>=0.16.0",
]

analytics = [
    "datadog>=0.47.0",
    "langsmith>=0.0.69",
]

[project.urls]
Homepage = "https://github.com/your-org/email-router"
Repository = "https://github.com/your-org/email-router.git"
Documentation = "https://docs.email-router.ai"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false 